fix: display existing values on edit
