+++ packages/javascript/bh-shared-ui/src/components/UpsertOIDCProviderDialog.tsx
@@ -14,42 +14,17 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import { Button } from '@bloodhoundenterprise/doodleui';
-import { Alert, Dialog, DialogTitle, DialogContent, DialogActions, Grid, TextField } from '@mui/material';
-import { FC } from 'react';
-import { Controller, useForm } from 'react-hook-form';
-import { UpsertOIDCProviderRequest } from 'js-client-library';
+import { Dialog, DialogTitle } from '@mui/material';
+import { SSOProvider, UpsertOIDCProviderRequest } from 'js-client-library';
+import UpsertOIDCProviderForm from './UpsertOIDCProviderForm';
 
-const UpsertOIDCProviderDialog: FC<{
+const UpsertSAMLProviderDialog: React.FC<{
     open: boolean;
     error?: string;
-    isUpdate: boolean;
+    oldSSOProvider?: SSOProvider;
     onClose: () => void;
     onSubmit: (data: UpsertOIDCProviderRequest) => void;
-}> = ({ open, error, isUpdate, onClose, onSubmit: _onSubmit }) => {
-    const {
-        control,
-        handleSubmit,
-        reset,
-        formState: { errors },
-    } = useForm<UpsertOIDCProviderRequest>({
-        defaultValues: {
-            name: '',
-            client_id: '',
-            issuer: '',
-        },
-    });
-
-    const onSubmit = (data: UpsertOIDCProviderRequest) => {
-        _onSubmit(data);
-        reset();
-    };
-
-    const handleClose = () => {
-        onClose();
-        reset();
-    };
-
+}> = ({ open, error, oldSSOProvider, onClose, onSubmit }) => {
     return (
         <Dialog
             open={open}
@@ -60,101 +35,15 @@ const UpsertOIDCProviderDialog: FC<{
                 // @ts-ignore
                 'data-testid': 'create-oidc-provider-dialog',
             }}>
-            <DialogTitle>{isUpdate ? 'Edit' : 'Create'} OIDC Provider</DialogTitle>
-            <form onSubmit={handleSubmit(onSubmit)}>
-                <DialogContent>
-                    <Grid container spacing={2}>
-                        <Grid item xs={12}>
-                            <Controller
-                                control={control}
-                                name='name'
-                                rules={{
-                                    required: 'OIDC Provider Name is required',
-                                    pattern: {
-                                        value: /^[A-z0-9 ]+$/,
-                                        message: 'OIDC Provider Name must be alphanumeric.',
-                                    },
-                                }}
-                                render={({ field }) => (
-                                    <TextField
-                                        {...field}
-                                        id={'name'}
-                                        variant='standard'
-                                        fullWidth
-                                        name='name'
-                                        label='OIDC Provider Name'
-                                        error={!!errors.name}
-                                        helperText={
-                                            errors.name?.message || 'Choose a name for your OIDC Provider configuration'
-                                        }
-                                    />
-                                )}
-                            />
-                        </Grid>
-                        <Grid item xs={12}>
-                            <Controller
-                                control={control}
-                                name='client_id'
-                                rules={{
-                                    required: 'Client ID is required',
-                                }}
-                                render={({ field }) => (
-                                    <TextField
-                                        {...field}
-                                        id={'clientId'}
-                                        variant='standard'
-                                        fullWidth
-                                        name='clientId'
-                                        label='Client ID'
-                                        error={!!errors.client_id}
-                                        helperText={errors.client_id?.message || 'OIDC Provider Client ID'}
-                                    />
-                                )}
-                            />
-                        </Grid>
-                        <Grid item xs={12}>
-                            <Controller
-                                control={control}
-                                name='issuer'
-                                rules={{
-                                    required: 'Issuer is required',
-                                }}
-                                render={({ field }) => (
-                                    <TextField
-                                        {...field}
-                                        id={'issuer'}
-                                        variant='standard'
-                                        fullWidth
-                                        name='issuer'
-                                        label='Issuer'
-                                        error={!!errors.issuer}
-                                        helperText={errors.issuer?.message || 'OIDC Issuer'}
-                                    />
-                                )}
-                            />
-                        </Grid>
-                        {error && (
-                            <Grid item xs={12}>
-                                <Alert severity='error'>{error}</Alert>
-                            </Grid>
-                        )}
-                    </Grid>
-                </DialogContent>
-                <DialogActions>
-                    <Button
-                        type='button'
-                        variant='tertiary'
-                        onClick={handleClose}
-                        data-testid='create-oidc-provider-dialog_button-close'>
-                        Cancel
-                    </Button>
-                    <Button data-testid='create-oidc-provider-dialog_button-save' type='submit'>
-                        Submit
-                    </Button>
-                </DialogActions>
-            </form>
+            <DialogTitle>{oldSSOProvider ? 'Edit' : 'Create'} OIDC Provider</DialogTitle>
+            <UpsertOIDCProviderForm
+                error={error}
+                onClose={onClose}
+                oldSSOProvider={oldSSOProvider}
+                onSubmit={onSubmit}
+            />
         </Dialog>
     );
 };
 
-export default UpsertOIDCProviderDialog;
+export default UpsertSAMLProviderDialog;
